Rename Connection to Protocol.
This makes the following naming possible:

* Connection = TCP/TLS connection + pointer to protocol; concerned with
  opening the network connection, moving bytes, and closing it.
* Protocol = Sans-I/O protocol; concerned with parsing and serializing
  messages and with the state of the connection.

Previously, these names were reversed in the sockets & threads branch.

The previous choice was influenced by the legacy implementation using
"protocol" to describe the two layers together, itself influenced by
asyncio using the same word.
