+++ tests/test_client.py
@@ -3,11 +3,11 @@ import unittest
 import unittest.mock
 
 from websockets.client import *
-from websockets.connection import CONNECTING, OPEN
 from websockets.datastructures import Headers
 from websockets.exceptions import InvalidHandshake, InvalidHeader
 from websockets.frames import OP_TEXT, Frame
 from websockets.http11 import Request, Response
+from websockets.protocol import CONNECTING, OPEN
 from websockets.uri import parse_uri
 from websockets.utils import accept_key
 
@@ -18,13 +18,13 @@ from .extensions.utils import (
     Rsv2Extension,
 )
 from .test_utils import ACCEPT, KEY
-from .utils import DATE
+from .utils import DATE, DeprecationTestCase
 
 
 class ConnectTests(unittest.TestCase):
     def test_send_connect(self):
         with unittest.mock.patch("websockets.client.generate_key", return_value=KEY):
-            client = ClientConnection(parse_uri("wss://example.com/test"))
+            client = ClientProtocol(parse_uri("wss://example.com/test"))
         request = client.connect()
         self.assertIsInstance(request, Request)
         client.send_request(request)
@@ -44,7 +44,7 @@ class ConnectTests(unittest.TestCase):
 
     def test_connect_request(self):
         with unittest.mock.patch("websockets.client.generate_key", return_value=KEY):
-            client = ClientConnection(parse_uri("wss://example.com/test"))
+            client = ClientProtocol(parse_uri("wss://example.com/test"))
         request = client.connect()
         self.assertEqual(request.path, "/test")
         self.assertEqual(
@@ -61,7 +61,7 @@ class ConnectTests(unittest.TestCase):
         )
 
     def test_path(self):
-        client = ClientConnection(parse_uri("wss://example.com/endpoint?test=1"))
+        client = ClientProtocol(parse_uri("wss://example.com/endpoint?test=1"))
         request = client.connect()
 
         self.assertEqual(request.path, "/endpoint?test=1")
@@ -76,19 +76,19 @@ class ConnectTests(unittest.TestCase):
             ("wss://example.com:8443/", "example.com:8443"),
         ]:
             with self.subTest(uri=uri):
-                client = ClientConnection(parse_uri(uri))
+                client = ClientProtocol(parse_uri(uri))
                 request = client.connect()
 
                 self.assertEqual(request.headers["Host"], host)
 
     def test_user_info(self):
-        client = ClientConnection(parse_uri("wss://hello:iloveyou@example.com/"))
+        client = ClientProtocol(parse_uri("wss://hello:iloveyou@example.com/"))
         request = client.connect()
 
         self.assertEqual(request.headers["Authorization"], "Basic aGVsbG86aWxvdmV5b3U=")
 
     def test_origin(self):
-        client = ClientConnection(
+        client = ClientProtocol(
             parse_uri("wss://example.com/"),
             origin="https://example.com",
         )
@@ -97,7 +97,7 @@ class ConnectTests(unittest.TestCase):
         self.assertEqual(request.headers["Origin"], "https://example.com")
 
     def test_extensions(self):
-        client = ClientConnection(
+        client = ClientProtocol(
             parse_uri("wss://example.com/"),
             extensions=[ClientOpExtensionFactory()],
         )
@@ -106,7 +106,7 @@ class ConnectTests(unittest.TestCase):
         self.assertEqual(request.headers["Sec-WebSocket-Extensions"], "x-op; op")
 
     def test_subprotocols(self):
-        client = ClientConnection(
+        client = ClientProtocol(
             parse_uri("wss://example.com/"),
             subprotocols=["chat"],
         )
@@ -118,7 +118,7 @@ class ConnectTests(unittest.TestCase):
 class AcceptRejectTests(unittest.TestCase):
     def test_receive_accept(self):
         with unittest.mock.patch("websockets.client.generate_key", return_value=KEY):
-            client = ClientConnection(parse_uri("ws://example.com/test"))
+            client = ClientProtocol(parse_uri("ws://example.com/test"))
         client.connect()
         client.receive_data(
             (
@@ -138,7 +138,7 @@ class AcceptRejectTests(unittest.TestCase):
 
     def test_receive_reject(self):
         with unittest.mock.patch("websockets.client.generate_key", return_value=KEY):
-            client = ClientConnection(parse_uri("ws://example.com/test"))
+            client = ClientProtocol(parse_uri("ws://example.com/test"))
         client.connect()
         client.receive_data(
             (
@@ -159,7 +159,7 @@ class AcceptRejectTests(unittest.TestCase):
 
     def test_accept_response(self):
         with unittest.mock.patch("websockets.client.generate_key", return_value=KEY):
-            client = ClientConnection(parse_uri("ws://example.com/test"))
+            client = ClientProtocol(parse_uri("ws://example.com/test"))
         client.connect()
         client.receive_data(
             (
@@ -189,7 +189,7 @@ class AcceptRejectTests(unittest.TestCase):
 
     def test_reject_response(self):
         with unittest.mock.patch("websockets.client.generate_key", return_value=KEY):
-            client = ClientConnection(parse_uri("ws://example.com/test"))
+            client = ClientProtocol(parse_uri("ws://example.com/test"))
         client.connect()
         client.receive_data(
             (
@@ -235,7 +235,7 @@ class AcceptRejectTests(unittest.TestCase):
         )
 
     def test_basic(self):
-        client = ClientConnection(parse_uri("wss://example.com/"))
+        client = ClientProtocol(parse_uri("wss://example.com/"))
         response = self.make_accept_response(client)
         client.receive_data(response.serialize())
         [response] = client.events_received()
@@ -243,7 +243,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(client.state, OPEN)
 
     def test_missing_connection(self):
-        client = ClientConnection(parse_uri("wss://example.com/"))
+        client = ClientProtocol(parse_uri("wss://example.com/"))
         response = self.make_accept_response(client)
         del response.headers["Connection"]
         client.receive_data(response.serialize())
@@ -255,7 +255,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(str(raised.exception), "missing Connection header")
 
     def test_invalid_connection(self):
-        client = ClientConnection(parse_uri("wss://example.com/"))
+        client = ClientProtocol(parse_uri("wss://example.com/"))
         response = self.make_accept_response(client)
         del response.headers["Connection"]
         response.headers["Connection"] = "close"
@@ -268,7 +268,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(str(raised.exception), "invalid Connection header: close")
 
     def test_missing_upgrade(self):
-        client = ClientConnection(parse_uri("wss://example.com/"))
+        client = ClientProtocol(parse_uri("wss://example.com/"))
         response = self.make_accept_response(client)
         del response.headers["Upgrade"]
         client.receive_data(response.serialize())
@@ -280,7 +280,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(str(raised.exception), "missing Upgrade header")
 
     def test_invalid_upgrade(self):
-        client = ClientConnection(parse_uri("wss://example.com/"))
+        client = ClientProtocol(parse_uri("wss://example.com/"))
         response = self.make_accept_response(client)
         del response.headers["Upgrade"]
         response.headers["Upgrade"] = "h2c"
@@ -293,7 +293,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(str(raised.exception), "invalid Upgrade header: h2c")
 
     def test_missing_accept(self):
-        client = ClientConnection(parse_uri("wss://example.com/"))
+        client = ClientProtocol(parse_uri("wss://example.com/"))
         response = self.make_accept_response(client)
         del response.headers["Sec-WebSocket-Accept"]
         client.receive_data(response.serialize())
@@ -305,7 +305,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(str(raised.exception), "missing Sec-WebSocket-Accept header")
 
     def test_multiple_accept(self):
-        client = ClientConnection(parse_uri("wss://example.com/"))
+        client = ClientProtocol(parse_uri("wss://example.com/"))
         response = self.make_accept_response(client)
         response.headers["Sec-WebSocket-Accept"] = ACCEPT
         client.receive_data(response.serialize())
@@ -321,7 +321,7 @@ class AcceptRejectTests(unittest.TestCase):
         )
 
     def test_invalid_accept(self):
-        client = ClientConnection(parse_uri("wss://example.com/"))
+        client = ClientProtocol(parse_uri("wss://example.com/"))
         response = self.make_accept_response(client)
         del response.headers["Sec-WebSocket-Accept"]
         response.headers["Sec-WebSocket-Accept"] = ACCEPT
@@ -336,7 +336,7 @@ class AcceptRejectTests(unittest.TestCase):
         )
 
     def test_no_extensions(self):
-        client = ClientConnection(parse_uri("wss://example.com/"))
+        client = ClientProtocol(parse_uri("wss://example.com/"))
         response = self.make_accept_response(client)
         client.receive_data(response.serialize())
         [response] = client.events_received()
@@ -345,7 +345,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(client.extensions, [])
 
     def test_no_extension(self):
-        client = ClientConnection(
+        client = ClientProtocol(
             parse_uri("wss://example.com/"),
             extensions=[ClientOpExtensionFactory()],
         )
@@ -358,7 +358,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(client.extensions, [OpExtension()])
 
     def test_extension(self):
-        client = ClientConnection(
+        client = ClientProtocol(
             parse_uri("wss://example.com/"),
             extensions=[ClientRsv2ExtensionFactory()],
         )
@@ -371,7 +371,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(client.extensions, [Rsv2Extension()])
 
     def test_unexpected_extension(self):
-        client = ClientConnection(parse_uri("wss://example.com/"))
+        client = ClientProtocol(parse_uri("wss://example.com/"))
         response = self.make_accept_response(client)
         response.headers["Sec-WebSocket-Extensions"] = "x-op; op"
         client.receive_data(response.serialize())
@@ -383,7 +383,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(str(raised.exception), "no extensions supported")
 
     def test_unsupported_extension(self):
-        client = ClientConnection(
+        client = ClientProtocol(
             parse_uri("wss://example.com/"),
             extensions=[ClientRsv2ExtensionFactory()],
         )
@@ -401,7 +401,7 @@ class AcceptRejectTests(unittest.TestCase):
         )
 
     def test_supported_extension_parameters(self):
-        client = ClientConnection(
+        client = ClientProtocol(
             parse_uri("wss://example.com/"),
             extensions=[ClientOpExtensionFactory("this")],
         )
@@ -414,7 +414,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(client.extensions, [OpExtension("this")])
 
     def test_unsupported_extension_parameters(self):
-        client = ClientConnection(
+        client = ClientProtocol(
             parse_uri("wss://example.com/"),
             extensions=[ClientOpExtensionFactory("this")],
         )
@@ -432,7 +432,7 @@ class AcceptRejectTests(unittest.TestCase):
         )
 
     def test_multiple_supported_extension_parameters(self):
-        client = ClientConnection(
+        client = ClientProtocol(
             parse_uri("wss://example.com/"),
             extensions=[
                 ClientOpExtensionFactory("this"),
@@ -448,7 +448,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(client.extensions, [OpExtension("that")])
 
     def test_multiple_extensions(self):
-        client = ClientConnection(
+        client = ClientProtocol(
             parse_uri("wss://example.com/"),
             extensions=[ClientOpExtensionFactory(), ClientRsv2ExtensionFactory()],
         )
@@ -462,7 +462,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(client.extensions, [OpExtension(), Rsv2Extension()])
 
     def test_multiple_extensions_order(self):
-        client = ClientConnection(
+        client = ClientProtocol(
             parse_uri("wss://example.com/"),
             extensions=[ClientOpExtensionFactory(), ClientRsv2ExtensionFactory()],
         )
@@ -476,7 +476,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(client.extensions, [Rsv2Extension(), OpExtension()])
 
     def test_no_subprotocols(self):
-        client = ClientConnection(parse_uri("wss://example.com/"))
+        client = ClientProtocol(parse_uri("wss://example.com/"))
         response = self.make_accept_response(client)
         client.receive_data(response.serialize())
         [response] = client.events_received()
@@ -485,9 +485,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertIsNone(client.subprotocol)
 
     def test_no_subprotocol(self):
-        client = ClientConnection(
-            parse_uri("wss://example.com/"), subprotocols=["chat"]
-        )
+        client = ClientProtocol(parse_uri("wss://example.com/"), subprotocols=["chat"])
         response = self.make_accept_response(client)
         client.receive_data(response.serialize())
         [response] = client.events_received()
@@ -496,9 +494,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertIsNone(client.subprotocol)
 
     def test_subprotocol(self):
-        client = ClientConnection(
-            parse_uri("wss://example.com/"), subprotocols=["chat"]
-        )
+        client = ClientProtocol(parse_uri("wss://example.com/"), subprotocols=["chat"])
         response = self.make_accept_response(client)
         response.headers["Sec-WebSocket-Protocol"] = "chat"
         client.receive_data(response.serialize())
@@ -508,7 +504,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(client.subprotocol, "chat")
 
     def test_unexpected_subprotocol(self):
-        client = ClientConnection(parse_uri("wss://example.com/"))
+        client = ClientProtocol(parse_uri("wss://example.com/"))
         response = self.make_accept_response(client)
         response.headers["Sec-WebSocket-Protocol"] = "chat"
         client.receive_data(response.serialize())
@@ -520,7 +516,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(str(raised.exception), "no subprotocols supported")
 
     def test_multiple_subprotocols(self):
-        client = ClientConnection(
+        client = ClientProtocol(
             parse_uri("wss://example.com/"),
             subprotocols=["superchat", "chat"],
         )
@@ -538,7 +534,7 @@ class AcceptRejectTests(unittest.TestCase):
         )
 
     def test_supported_subprotocol(self):
-        client = ClientConnection(
+        client = ClientProtocol(
             parse_uri("wss://example.com/"),
             subprotocols=["superchat", "chat"],
         )
@@ -551,7 +547,7 @@ class AcceptRejectTests(unittest.TestCase):
         self.assertEqual(client.subprotocol, "chat")
 
     def test_unsupported_subprotocol(self):
-        client = ClientConnection(
+        client = ClientProtocol(
             parse_uri("wss://example.com/"),
             subprotocols=["superchat", "chat"],
         )
@@ -568,7 +564,7 @@ class AcceptRejectTests(unittest.TestCase):
 
 class MiscTests(unittest.TestCase):
     def test_bypass_handshake(self):
-        client = ClientConnection(parse_uri("ws://example.com/test"), state=OPEN)
+        client = ClientProtocol(parse_uri("ws://example.com/test"), state=OPEN)
         client.receive_data(b"\x81\x06Hello!")
         [frame] = client.events_received()
         self.assertEqual(frame, Frame(OP_TEXT, b"Hello!"))
@@ -576,5 +572,17 @@ class MiscTests(unittest.TestCase):
     def test_custom_logger(self):
         logger = logging.getLogger("test")
         with self.assertLogs("test", logging.DEBUG) as logs:
-            ClientConnection(parse_uri("wss://example.com/test"), logger=logger)
+            ClientProtocol(parse_uri("wss://example.com/test"), logger=logger)
         self.assertEqual(len(logs.records), 1)
+
+
+class BackwardsCompatibilityTests(DeprecationTestCase):
+    def test_client_connection_class(self):
+        with self.assertDeprecationWarning(
+            "ClientConnection was renamed to ClientProtocol"
+        ):
+            from websockets.client import ClientConnection
+
+            client = ClientConnection("ws://localhost/")
+
+        self.assertIsInstance(client, ClientProtocol)
